# client/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Loom.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/institutions/partner": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "Country",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IncludeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SafePage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SafePageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInstitutionAdminResponsePage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PartnerInstitutionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInstitutionAdminResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/institutions/partner/{institutionId}": {
      "put": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PartnerInstitutionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInstitutionAdminResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/institutions/partner/{institutionId}/restore": {
      "patch": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/institutions/partner/{institutionId}/courses": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Semester",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ExchangeSemester"
                }
              ]
            }
          },
          {
            "name": "Level",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/StudyProgramLevel"
                }
              ]
            }
          },
          {
            "name": "IncludeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SafePage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SafePageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseResponsePage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "institutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PartnerCourseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/institutions/partner/courses/{courseId}": {
      "put": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PartnerCourseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/institutions/partner/courses/{courseId}/restore": {
      "patch": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/institutions/partner/courses/merge": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/MergePartnerCoursesRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/institutions/partner/courses/{courseId}/usage": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseUsageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExchangeSemester": {
        "enum": [
          "Winter",
          "Summer",
          "Both"
        ],
        "type": "string"
      },
...
      "MergePartnerCoursesRequest": {
        "required": [
          "duplicateCourseIds",
          "primaryCourseId"
        ],
        "type": "object",
        "properties": {
          "primaryCourseId": {
            "type": "integer",
            "format": "int32"
          },
          "duplicateCourseIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseRequest": {
        "required": [
          "code",
          "ects",
          "level",
          "name",
          "semester"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameHr": {
            "type": "string",
            "nullable": true
          },
          "ects": {
            "type": "number",
            "format": "double"
          },
          "semester": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "lecturesH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "auditoryH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "labH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseResponse": {
        "required": [
          "auditoryH",
          "code",
          "ects",
          "id",
          "isDeleted",
          "labH",
          "lecturesH",
          "level",
          "name",
          "nameHr",
          "semester",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameHr": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "ects": {
            "type": "number",
            "format": "double"
          },
          "lecturesH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "auditoryH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "labH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "semester": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExchangeSemester"
              }
            ]
          },
          "level": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StudyProgramLevel"
              }
            ]
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseResponsePage": {
        "required": [
          "hasDeleted",
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartnerCourseResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseUsageGroup": {
        "required": [
          "academicYears",
          "exchangeCount",
          "isCourseGroup",
          "profileName",
          "programName",
          "recognizedAsIsvuCode",
          "recognizedAsName",
          "totalAwardedEcts"
        ],
        "type": "object",
        "properties": {
          "programName": {
            "type": "string"
          },
          "profileName": {
            "type": "string"
          },
          "recognizedAsIsvuCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recognizedAsName": {
            "type": "string"
          },
          "isCourseGroup": {
            "type": "boolean"
          },
          "exchangeCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalAwardedEcts": {
            "type": "number",
            "format": "double"
          },
          "academicYears": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseUsageResponse": {
        "required": [
          "exchangeCount",
          "groups"
        ],
        "type": "object",
        "properties": {
          "exchangeCount": {
            "type": "integer",
            "format": "int32"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartnerCourseUsageGroup"
            }
          }
        },
        "additionalProperties": false
      },
      "PartnerInstitutionAdminResponse": {
        "required": [
          "city",
          "country",
          "courseCount",
          "erasmusCode",
          "id",
          "isDeleted",
          "name",
          "nameHr"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "nameHr": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "erasmusCode": {
            "type": "string",
            "nullable": true
          },
          "courseCount": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PartnerInstitutionAdminResponsePage": {
        "required": [
          "hasDeleted",
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartnerInstitutionAdminResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PartnerInstitutionRequest": {
        "required": [
          "country",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "nameHr": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "erasmusCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StudyProgramLevel": {
        "enum": [
          "Undergraduate",
          "Graduate",
          "Postgraduate"
        ],
        "type": "string"
      },
...
    }
  },
...
}